+++ crates/nostr-relay-builder/examples/mock.rs
@@ -1,0 +1,21 @@
+// Copyright (c) 2022-2023 Yuki Kishimoto
+// Copyright (c) 2023-2024 Rust Nostr Developers
+// Distributed under the MIT software license
+
+use std::time::Duration;
+
+use nostr_relay_builder::prelude::*;
+
+#[tokio::main]
+async fn main() -> Result<()> {
+    tracing_subscriber::fmt::init();
+
+    let relay = MockRelay::run().await?;
+
+    println!("Url: {}", relay.url());
+
+    // Keep up the program
+    loop {
+        tokio::time::sleep(Duration::from_secs(60)).await;
+    }
+}