+++ crates/nostr-relay-builder/src/lib.rs
@@ -1,0 +1,15 @@
+// Copyright (c) 2022-2023 Yuki Kishimoto
+// Copyright (c) 2023-2024 Rust Nostr Developers
+// Distributed under the MIT software license
+
+//! Nostr Relay Builder and Mock Relay for tests
+
+#![forbid(unsafe_code)]
+#![warn(missing_docs)]
+#![warn(rustdoc::bare_urls)]
+#![warn(clippy::large_futures)]
+
+// pub struct RelayBuilder {}
+
+pub mod mock;
+pub mod prelude;