+++ crates/nostr-relay-builder/src/prelude.rs
@@ -1,0 +1,16 @@
+// Copyright (c) 2022-2023 Yuki Kishimoto
+// Copyright (c) 2023-2024 Rust Nostr Developers
+// Distributed under the MIT software license
+
+//! Prelude
+
+#![allow(unknown_lints)]
+#![allow(ambiguous_glob_reexports)]
+#![doc(hidden)]
+
+// External crates
+pub use nostr::prelude::*;
+
+// Internal modules
+pub use crate::mock::{self, *};
+pub use crate::*;