+++ CHANGELOG.md
@@ -66,6 +66,7 @@
 * nostr: add `EventBuilder::pow` method ([Yuki Kishimoto])
 * nostr: add `TagKind::custom` constructor ([Yuki Kishimoto])
 * database: add `DatabaseHelper::fast_query` ([Yuki Kishimoto])
+* relay-builder: add `MockRelay` ([Yuki Kishimoto])
 * bindings(nostr): expose `as_pretty_json` for some structs ([Yuki Kishimoto])
 * bindings(sdk): expose `Client::fetch_metadata` ([Yuki Kishimoto])
 * ffi(nostr): expose `Kind::is_*` methods ([Yuki Kishimoto])